requirements.txt
Django==5.0.2 
python-dotenv==1.0.1 
requests==2.31.0 
Pillow
anthropic==0.25.1
whitenoise==6.6.0 
django-crispy-forms==2.1 
crispy-bootstrap5==2024.2
dj-database-url==2.1.0

